Set up PSR standard and test environments
