+++ .gitignore
@@ -1,0 +1,4 @@
+/vendor/
+/composer.lock
+/.phpunit.result.cache
+/.vscode
