+++ src/CommandObject.php
@@ -1,35 +1,33 @@
 <?php
 
+namespace Styoo4001\PhpDtoMapper;
+
+use JsonSerializable;
+use ReflectionClass;
+
 /**
- *
  * Class CommandObject
  */
-abstract class CommandObject implements RequestCommand, JsonSerializable
+abstract class CommandObject implements JsonSerializable, RequestCommand
 {
     protected array $commandObjectErrors = [];
+
     abstract protected function rule(): array;
 
-    /**
-     * @param CommandObjectValidator $validator
-     */
     public function validation(CommandObjectValidator $validator)
     {
         $validator->setRules($this->rule());
-        if (!$validator->run()) {
+        if (! $validator->run()) {
             $this->commandObjectErrors = array_merge($this->commandObjectErrors, $validator->getValidError());
         }
     }
 
-    /**
-     * @return bool
-     */
     public function hasErrors(): bool
     {
-        return !empty($this->commandObjectErrors);
+        return ! empty($this->commandObjectErrors);
     }
 
     /**
-     * @param string $key
      * @return false|mixed
      */
     public function getError(string $key)
@@ -37,17 +35,11 @@ abstract class CommandObject implements RequestCommand, JsonSerializable
         return $this->commandObjectErrors[$key] ?: false;
     }
 
-    /**
-     * @return array
-     */
     public function getErrors(): array
     {
         return $this->commandObjectErrors;
     }
 
-    /**
-     * @param string ...$errors
-     */
     public function setErrors(string ...$errors): void
     {
         foreach ($errors as $key => $error) {
@@ -57,9 +49,12 @@ abstract class CommandObject implements RequestCommand, JsonSerializable
 
     /**
      * Specify data which should be serialized to JSON
+     *
      * @link https://php.net/manual/en/jsonserializable.jsonserialize.php
+     *
      * @return mixed data which can be serialized by <b>json_encode</b>,
-     * which is a value of any type other than a resource.
+     *               which is a value of any type other than a resource.
+     *
      * @since 5.4
      */
     public function jsonSerialize(): array
@@ -75,6 +70,7 @@ abstract class CommandObject implements RequestCommand, JsonSerializable
             }
         }
         unset($json['commandObjectErrors']);
+
         return $json;
     }
 }
