+++ src/CommandObjectMapper.php
@@ -1,11 +1,12 @@
 <?php
 
+namespace Styoo4001\PhpDtoMapper;
+
 use Illuminate\Http\Request;
 
 /** Class CommandObjectMapper */
 class CommandObjectMapper
 {
-
     public static function mapping(RequestCommand $commandObject, array $mappingData)
     {
         $mapper = new DataTransferObjectMapper();
@@ -13,7 +14,7 @@ class CommandObjectMapper
         $validator->setData($mappingData);
         $commandObject->validation($validator);
 
-        if (!$commandObject->hasErrors()) {
+        if (! $commandObject->hasErrors()) {
             $mapper->mapping(array_merge($mappingData, $validator->getValidatedData()), $commandObject);
         }
 
@@ -21,6 +22,7 @@ class CommandObjectMapper
         if ($mapper->hasErrors()) {
             $commandObject->setErrors(...$mapper->getErrors());
         }
+
         // commandObject의 validator를 설정한다.
         return $commandObject;
     }
@@ -29,17 +31,19 @@ class CommandObjectMapper
     {
         $content = $request->getContent();
         //request 내용이 json content면 json_decode 한 후에 mapping.
-        if (!empty($request) && $request->isJson()) {
+        if (! empty($request) && $request->isJson()) {
             $jsonData = @json_decode($content, true);
 
             if (json_last_error() === JSON_ERROR_NONE) {
                 return $jsonData;
             } else {
                 parse_str($content, $mappingData);
+
                 return $mappingData;
             }
             // json요청이 아니면 $request->all() 로 받아온 데이터 전체를 mapping 시킨다.
         }
+
         return $request->all();
     }
 }
