+++ src/DataMappingException.php
@@ -1,6 +1,7 @@
 <?php
 
+namespace Styoo4001\PhpDtoMapper;
+
 class DataMappingException extends \Exception
 {
-
 }
