+++ src/Dto/Money.php
@@ -1,20 +1,19 @@
 <?php
 
+namespace Styoo4001\PhpDtoMapper\Dto;
+
 class Money
 {
     private int $amount;
+
     /**
      * Money constructor.
-     * @param int $amount
      */
     public function __construct(int $amount)
     {
         $this->amount = $amount;
     }
 
-    /**
-     * @return int
-     */
     public function getAmount(): int
     {
         return $this->amount;
