+++ src/ServiceProvider.php
@@ -1,8 +1,13 @@
 <?php
 
+namespace Styoo4001\PhpDtoMapper;
+
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Http\Request;
+
 class ServiceProvider
 {
-    // ... 
+    // ...
     public function boot(Application $app): void
     {
         $app->resolving(function ($object, $app) {
