+++ src/SomeValidator.php
@@ -1,9 +1,9 @@
 <?php
 
+namespace Styoo4001\PhpDtoMapper;
 
 class SomeValidator implements CommandObjectValidator
 {
-
     public function getValidator()
     {
         // TODO: Implement getValidator() method.
@@ -22,15 +22,18 @@ class SomeValidator implements CommandObjectValidator
     public function run(): bool
     {
         // TODO: Implement run() method.
+        return true;
     }
 
     public function getValidatedData(): array
     {
         // TODO: Implement getValidatedData() method.
+        return [];
     }
 
     public function getValidError(): array
     {
         // TODO: Implement getValidError() method.
+        return [];
     }
 }
