+++ tests/Fixtures/Money.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Tests\Fixtures;
+
+class Money
+{
+    private int $amount;
+
+    public function getAmount(): int
+    {
+        return $this->amount;
+    }
+}
