+++ tests/Fixtures/Skill.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Tests\Fixtures;
+
+class Skill
+{
+    private string $client;
+
+    private string $server;
+
+    public function getServer(): string
+    {
+        return $this->server;
+    }
+}
